refactor(domain): 객체지향프로그래밍 기반 도메인 아키텍처 설계
- 멘탈 모델의 행위자를 service로 분리하여 LottoGame 책임 축소 (LottoMachine, LottoBroadcast, Statistics)
- 사물의 묶음을 일급 컬렉션으로 도입하여 의도를 드러내는 인터페이스 구성 (Lottos, Ranks)
- 추상화 수준에 따라 entity/collection/service 레이어 분할
