+++ docs/REFACTOR.md
@@ -2,6 +2,8 @@
 
 ## 멘탈 모델
 
+코드를 설계하기 전에, 현실의 로또가 어떻게 작동하는지 관찰했다.
+
 ### 현실의 로또 — 등장하는 것들
 
 ```text
@@ -42,23 +44,114 @@
 
 ### 핵심 관찰
 
+이 흐름을 보면 네 가지가 눈에 띈다:
+
 1. **"추첨 결과"는 하나의 덩어리** — 6개 번호 + 보너스를 따로 발표하지 않음
 2. **"발행"은 판매점의 행위** — 티켓 묶음이 스스로 자신을 발행하지 않음
 3. **"대조"와 "집계"는 나의 행위** — 티켓이 스스로 대조하지 않음
 4. **"나"는 행위의 주체이지 데이터 객체가 아님** — 오케스트레이터(LottoGame)가 "나"
 
+이 관찰이 이후 리팩토링의 판단 기준이 된다.
+
 ---
 
-## Phase 0: Rank 멘탈 모델 왜곡 해소
+## Phase 0: 뷰 책임이 도메인을 왜곡한 부분을 해소한다
 
-### 문제. Rank의 멘탈 모델이 구조에 의해 왜곡되어 있다
+핵심 관찰을 기준으로 기존 코드를 점검하니, 뷰(UI) 레이어의 요구사항이 도메인 모델의 인터페이스를 왜곡하고 있는 부분이 두 곳 발견되었다.
 
-Rank의 본질은 "등수 + 상금"이다. 그런데 판정 조건인 matchCount와 isBonusMatch까지 Rank의 인스턴스 속성으로 구겨넣다 보니, `NONE = new Rank(2, false, 0)`처럼 거짓말을 하거나 `FIRST`의 `isBonusMatch: false`처럼 의미 없는 값을 채워야 한다. 또한 이 값들은 뷰에서 "3개 일치" 같은 텍스트를 렌더링하기 위해 존재하는 것이므로, Rank가 도메인 책임이 아닌 뷰 책임까지 갖고 있는 상태다.
+### 문제 1. Rank가 뷰 표시용 데이터를 갖고 있다
+
+멘탈 모델에서 등수의 본질은 "등수 + 상금"이다. 그런데 기존 Rank는 판정 조건인 matchCount와 isBonusMatch까지 인스턴스 속성으로 갖고 있었다.
+
+이로 인해 `NONE = new Rank(2, false, 0)`처럼 거짓말을 하거나, `FIRST`의 `isBonusMatch: false`처럼 의미 없는 값을 채워야 했다. 이 값들은 뷰에서 "3개 일치" 텍스트를 렌더링하기 위해 존재하는 것이므로, Rank가 도메인 책임이 아닌 뷰 책임까지 갖고 있는 상태였다.
 
 **해결**: Rank에서 판정 조건을 제거하고 prize만 남긴다. 뷰 표시 책임은 뷰 레이어로 분리한다.
 
 - Rank 생성자를 `constructor(prize)`로 단순화
 - `from(matchCount, isBonusMatch)` switch 판정 로직은 유지 (입력값으로만 사용, 저장하지 않음)
 - RankStatistics.summary는 Rank 인스턴스를 직접 포함하도록 변경
-- matchCount/isBonusMatch 표시 데이터는 뷰 레이어가 직접
+- matchCount/isBonusMatch 표시 데이터는 뷰 레이어가 직접 관리
 
+### 문제 2. WinningLotto의 생성 인터페이스가 뷰 표시를 위해 분리되어 있다
+
+핵심 관찰 1 — "추첨 결과는 하나의 덩어리"인데, 기존 코드는 `setWinningNumbers()` + `setBonusNumber()` 2단계로 WinningLotto를 생성했다. 이는 콘솔/웹 뷰가 당첨 번호와 보너스 번호를 따로 입력받기 때문에, 뷰의 입력 단계에 맞춰 도메인 인터페이스를 구성한 것이다. Rank가 뷰 표시를 위해 matchCount를 가진 것과 같은 패턴으로, 그 결과 `#firstRankLotto`라는 불완전한 중간 상태가 도메인에 존재했다.
+
+**해결**: WinningLotto 생성을 `setWinningLotto(winningNumbers, bonusNumber)` 1단계로 통합하고, `#firstRankLotto` 중간 상태를 제거한다. 2단계 입력에 의한 중간 상태는 뷰 레이어가 관리한다.
+
+---
+
+## Phase 1: 완전한 OOP로 전환
+
+### 왜 OOP인가
+
+하나의 코드에 발행·조립·대조·집계가 모여 있으면, 코드를 읽는 사람은 "이 메서드가 무엇을 하는가"를 파악하기 위해 구현 세부사항을 전부 따라가야 한다. 책임을 객체 단위로 분리하면:
+
+- **가독성**: 메서드 이름이 의도를 드러내고, 구현 세부사항은 인터페이스 뒤에 캡슐화된다. `LottoMachine.issue(budget)`만 보면 "판매점이 발행한다"를 알 수 있고, 내부에서 난수를 어떻게 생성하고 배열을 어떻게 조립하는지는 알 필요가 없다.
+- **유지보수**: 변경의 영향 범위가 객체 단위로 격리된다. 발행 방식이 바뀌면 LottoMachine만 수정하면 되고, LottoGame은 건드릴 필요가 없다.
+
+기존 코드도 멘탈 모델의 **사물**에 착안하여 각각을 독립된 객체로 분리한 상태였다:
+
+| 사물          | 객체         | 부여한 책임                                                              |
+| ------------- | ------------ | ------------------------------------------------------------------------ |
+| 로또 번호(공) | LottoNumber  | 자신의 유효 범위(1~45)를 스스로 지킨다                                   |
+| 로또 티켓     | Lotto        | 6개 번호의 정렬·중복 검사, "이 번호를 포함하는가?"에 답한다              |
+| 구매 금액     | Budget       | 발행 가능 수량과 실제 비용을 스스로 계산한다                             |
+| 추첨 결과     | WinningLotto | 당첨 번호 + 보너스를 하나의 덩어리로 갖고, **티켓의 등수를 판정**한다    |
+| 등수          | Rank         | 상금을 알고, 판정 조건(matchCount, isBonusMatch)으로부터 자신을 결정한다 |
+
+> WinningLotto에 판정 로직(`getRank`)을 둔 이유: OOP는 현실을 그대로 복사하는 게 아닌, 영감을 받아 가상의 세계를 만드는 것이다. 현실에서 종이가 판정할 수 없더라도, 가상 세계에서는 책임을 가진 능동적인 객체를 설계할 수 있다.
+
+### 문제 1. 사물만 분리해서는 LottoGame의 책임이 줄지 않는다
+
+기존 코드는 멘탈 모델의 **사물**만 객체로 분리해뒀다. 자연히 LottoGame이 흐름 제어 외에도 발행·당첨 설정·대조·집계를 직접 수행하고 있었다.
+
+그러다 보니 LottoGame이 모든 사물의 조립 과정을 알아야 했다. 발행 하나만 해도 난수 생성 → Lotto 생성 → 배열 조립이라는 세부사항을 전부 안다. 사물 하나의 생성 방식이 바뀌면 LottoGame도 함께 바뀌는 구조다.
+
+| 책임      | LottoGame이 직접 수행하던 작업                          |
+| --------- | ------------------------------------------------------- |
+| 발행      | 난수 생성 → Lotto 생성 → 배열 조립                      |
+| 당첨 설정 | Lotto 생성 + LottoNumber 생성 + WinningLotto 조립       |
+| 대조      | `Lotto[]` 배열을 직접 순회하며 WinningLotto에 각각 비교 |
+| 집계      | `Rank[]` 배열을 직접 순회하며 통계 계산                 |
+
+**해결 1: 멘탈 모델의 행위자를 객체로 분리하여, 조립 책임을 LottoGame에서 덜어낸다.**
+
+| 행위자          | 객체           | 부여한 책임                                                |
+| --------------- | -------------- | ---------------------------------------------------------- |
+| 판매점/로또기계 | LottoMachine   | 돈을 받고 티켓을 발행한다                                  |
+| 추첨 방송       | LottoBroadcast | 번호 배열 + 보너스 번호로 WinningLotto를 조립하여 발표한다 |
+| 집계하는 나     | Statistics     | 등수 목록과 구매 비용으로 통계/수익률을 계산한다           |
+
+멘탈 모델에서 행위자가 하던 조립을 각 객체에 맡기면, LottoGame은 조립 세부사항을 모른 채 "무엇을 해라"만 말할 수 있다. `generateLottoNumbers`는 `Lotto.DIGITS`, `LottoNumber.LOWER_BOUND/UPPER_BOUND` 등 사물의 지식에 전적으로 의존하므로, "자기 자신을 랜덤으로 만드는 법"은 Lotto의 자기 지식으로 보고 `Lotto.random()` 팩토리로 캡슐화했다.
+
+이 시점에서 LottoGame에는 흐름 제어 + 배열(`Lotto[]`, `Rank[]`) 관리의 책임이 남는다.
+
+**해결 2: 사물의 묶음을 일급 컬렉션으로 분리하여, 배열 관리 책임까지 덜어낸다.**
+
+행위자를 분리한 후에도 LottoGame에는 배열(`Lotto[]`, `Rank[]`)을 직접 순회하는 책임이 남아 있다. 일급 컬렉션으로 감싸면 배열 순회가 `lottos.determineRanks(winningLotto)`라는 메시지가 된다 — 명령형(how)에서 선언형(what)으로, 의도를 드러내는 코드가 된다.
+
+| 묶음              | 객체   | 부여한 책임                                                      |
+| ----------------- | ------ | ---------------------------------------------------------------- |
+| 로또 티켓 여러 장 | Lottos | "내 티켓들을 대조하면 등수가 뭐야?" — `determineRanks(winningLotto)` |
+| 등수 목록         | Ranks  | "등수별로 몇 개야?", "총 상금이 얼마야?" — `countByRank()`, `totalPrize()` |
+
+Lottos에서 `issue(budget)`를 제거한 것은, 핵심 관찰 2 — "발행은 판매점의 행위"이지, 티켓 묶음이 스스로를 발행하는 것이 아니기 때문이다. `determineRanks`가 `Ranks`를 반환하도록 바꾼 것은, 일급 컬렉션에서 나온 결과도 일급 컬렉션이어야 하기 때문이다.
+
+이제 LottoGame에는 흐름 제어만 남는다:
+
+1. `LottoMachine.issue(budget)` — 판매점에 돈을 내고 티켓을 받는다
+2. `LottoBroadcast.announce(numbers, bonusNumber)` — 추첨 방송의 결과를 받는다
+3. `lottos.determineRanks(winningLotto)` — 내 티켓들을 추첨 결과에 대조한다
+4. `Statistics.from(ranks, totalCost)` — 결과를 집계한다
+
+### 문제 2. 도메인 객체가 너무 많다
+
+사물, 행위자, 일급 컬렉션을 도입하면서 도메인 객체 수가 늘어났다. LottoNumber, Lotto, Budget, WinningLotto, Rank, Lottos, Ranks, LottoMachine, LottoBroadcast, Statistics — 이들이 모두 같은 `domain/` 아래 나열되면, 어떤 객체가 사물이고 어떤 객체가 행위자인지 구조만으로는 알 수 없다.
+
+**해결: 추상화 수준에 따라 entity, collection, service 레이어로 나눈다**
+
+- **entity** — 스스로의 유효성을 지키고 질문에 답하는 사물: LottoNumber, Lotto, Budget, WinningLotto, Rank
+- **collection** — 내부 배열을 노출하지 않고 묶음 단위의 행위를 제공하는 일급 컬렉션: Lottos, Ranks
+- **service** — 다른 객체들을 조합하여 행위를 수행하는 행위자: LottoMachine, LottoBroadcast, Statistics
+
+폴더 이름이 곧 추상화 수준을 말해주므로, 새로운 개발자가 코드베이스에 진입할 때 "이 객체는 데이터인가, 컬렉션인가, 행위자인가?"를 폴더 구조만으로 파악할 수 있고, 새 객체를 추가할 때도 어디에 놓아야 할지 고민할 필요가 없다.
