+++ src/js/domain/entity/LottoNumber/LottoNumber.js
@@ -1,0 +1,39 @@
+import { isNumber } from '../../../utils.js';
+import { ValueNotNumberError, ValueNotIntegerError, ValueOutOfRangeError } from './errors.js';
+
+export default class LottoNumber {
+    #value;
+
+    static LOWER_BOUND = 1;
+    static UPPER_BOUND = 45;
+
+    static #LOTTO_NUMBERS = Array.from(
+        { length: LottoNumber.UPPER_BOUND - LottoNumber.LOWER_BOUND + 1 },
+        (_, index) => new LottoNumber(index + 1),
+    );
+
+    static of(value) {
+        LottoNumber.#validate(value);
+        return LottoNumber.#LOTTO_NUMBERS[value - 1];
+    }
+
+    static #isInRange(target) {
+        return LottoNumber.LOWER_BOUND <= target && target <= LottoNumber.UPPER_BOUND;
+    }
+
+    static #validate(value) {
+        if (!isNumber(value)) throw new ValueNotNumberError();
+        if (!Number.isInteger(value)) throw new ValueNotIntegerError();
+        if (!LottoNumber.#isInRange(value))
+            throw new ValueOutOfRangeError(LottoNumber.LOWER_BOUND, LottoNumber.UPPER_BOUND);
+    }
+
+    constructor(value) {
+        LottoNumber.#validate(value);
+        this.#value = value;
+    }
+
+    get value() {
+        return this.#value;
+    }
+}
