+++ src/js/domain/entity/WinningLotto/WinningLotto.js
